backend/requirements.txt
fastapi[standard]==0.115.12
sqlmodel==0.0.24
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-slugify==8.0.4
text-unidecode==1.3
alembic==1.16.1
cloudinary==1.44.0
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
google-auth[requests]==2.40.3
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.0
uvicorn[standard]==0.30.6
slowapi==0.1.9
redis==5.0.1
psutil==5.9.8
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
